# Automotive_Shop_app/application/static/swagger.yaml
swagger: "2.0"
info:
  title: Automotive Repair Shop API
  version: "1.0"
  description: This API allows you to manage all facets of an Automive Repair Shop.
host: 127.0.0.1:5000
basePath: /
schemes:
  - http
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header
paths:
  /customers:
    post:
      tags:
        - Customers
      summary: Create a new customer
      description: Create a customer with the provided details.
      parameters:
        - in: body
          name: body
          description: Customer details for creating a new customer.
          required: true
          schema:
            $ref: "#/definitions/CustomerInput"
      responses:
        "201":
          description: Customer created successfully.
          schema:
            $ref: "#/definitions/Customer"
        "400":
          description: Invalid input.
    get:
      tags:
        - Customers
      summary: Retrieve all customers
      responses:
        "200":
          description: A list of customers.
          schema:
            type: array
            items:
              $ref: "#/definitions/Customer"
    put:
      tags:
        - Customers
      summary: Update customer information (requires authentication)
      security:
        - BearerAuth: []
      parameters:
        - in: body
          name: body
          description: Customer fields to update.
          required: true
          schema:
            $ref: "#/definitions/CustomerUpdate"
      responses:
        "200":
          description: Customer updated successfully.
          schema:
            $ref: "#/definitions/Customer"
        "400":
          description: Invalid customer ID or input.
    delete:
      tags:
        - Customers
      summary: Delete a customer (requires authentication & login)
      security:
        - BearerAuth: []
      parameters:
        - in: body
          name: body
          descripition: Information to update a customer.    
          required: true          
      responses:
        "200":
          description: Customer deleted successfully.
          schema:
            type: object
            properties:
              message:
                type: string
        "400":
          description: Invalid customer ID.
  /customers/login:
    post:
      tags:
        - Customers
      summary: Customer login
      parameters:
        - in: body
          name: body
          description: Login credentials.
          required: true
          schema:
            $ref: "#/definitions/LoginInput"
      responses:
        "200":
          description: Successfully logged in.
          schema:
            type: object
            properties:
              status:
                type: string
              message:
                type: string
              auth_token:
                type: string
        "400":
          description: Invalid payload.
        "401":
          description: Invalid username or password.
  /customers/{phone_number}:
    get:
      tags:
        - Customers
      summary: Get a customer by phone number
      parameters:
        - in: path
          name: phone_number
          type: string
          required: true
          description: The customer's phone number.
      responses:
        "200":
          description: Customer retrieved successfully.
          schema:
            $ref: "#/definitions/Customer"
        "400":
          description: Invalid customer id.
  /customers/my_tickets:
    get:
      tags:
        - Customers
      summary: Get service tickets for the authenticated customer (requires authentication)
      security:
        - BearerAuth: []
      responses:
        "200":
          description: A list of service tickets for the customer.
          schema:
            type: array
            items:
              $ref: "#/definitions/ServiceTicket"
        "400":
          description: Invalid customer id or no tickets found.
...
definitions:
  Customer:
    type: object
    properties:
      name:
        type: string
      phone_number:
        type: string
      email:
        type: string
    required:
      - name
      - phone_number
      - email
  CustomerInput:
    type: object
    properties:
      name:
        type: string
      phone_number:
        type: string
      password:
        type: string
      email:
        type: string
    required:
      - name
      - phone_number
      - password
      - email
  CustomerUpdate:
    type: object
    properties:
      name:
        type: string
      phone_number:
        type: string
    description: Optional fields to update.
  LoginInput:
    type: object
    properties:
      email:
        type: string
      password:
        type: string
    required:
      - email
      - password
...
  ServiceTicket:
    type: object
    properties:
      ticket_id:
        type: integer
      customer_phone:
        type: string
      vin:
        type: string
      services:
        type: string
    required:
      - ticket_id
      - customer_phone
      - services
...